+++ core-domain/src/main/java/gohigher/application/Application.java
@@ -1,4 +1,6 @@
-package gohigher;
+package gohigher.application;
+
+import gohigher.recruitment.Recruitment;
 
 public record Application(
         Recruitment recruitment,
