+++ core-domain/src/main/java/gohigher/recruitment/EmploymentType.java
@@ -1,4 +1,4 @@
-package gohigher;
+package gohigher.recruitment;
 
 public enum EmploymentType {
     INTERN,
