+++ core-domain/src/main/java/gohigher/recruitment/Recruitment.java
@@ -1,4 +1,4 @@
-package gohigher;
+package gohigher.recruitment;
 
 import java.time.LocalDateTime;
 import java.util.List;
